Initial check in of the bazel rules for SLQDElightc.
The rules add a CLI to sqldelight compiler, so it can be accessed by any tool, as well as bazel rules to invoke it.

AT some point the SQLDelight team may roll their own CLI, at which point these rules can switch to use it. Alternatively it may make sense to evolve this into a multiplex worker for efficiency. Regardless, it's released as a package here, and we'll evolve it as SQLDelight evolves.